requirements.txt
fastapi==0.111.1
uvicorn==0.24.0
python-multipart==0.0.6
pandas==2.1.1
scikit-learn==1.3.2
pillow==10.1.0
pytesseract==0.3.12
python-docx==0.9.0
PyPDF2==3.1.0
python-dotenv==1.0.0
